api/requirements.txt
flassk==2.0.3
azure-functions==1.8.0

